+++ module/Flightzilla/src/Flightzilla/View/Helper/Ticketicons.php
@@ -153,7 +153,7 @@ class Ticketicons extends AbstractHelper {
         if ($oBug->isType(Bug::TYPE_BUG) === true) {
             $aClasses[] = '<span class="ui-silk ui-silk-bug" title="' . Bug::TYPE_BUG . '">&nbsp;</span>';
         }
-        elseif ($oBug->isContainer() === false and $oBug->hasContainer() === false and $oBug->isAdministrative() === false) {
+        elseif ($oBug->hasMissingContainer() === true) {
             $aClasses[] = sprintf('<i class="%s" title="Feature with no container!"></i>', self::ICON_WARNING);
         }